readme update, message scaladooc, externalizing MessageSerializer
